"working" attempt functionality

+++ Models/Entities/Attempt.cs
@@ -8,7 +8,7 @@ namespace maturigo.Models.Entities
         public string Id { get; set; }
         public string UserId { get; set; }
         public string ExamId { get; set; }
-        public float Score { get; set; }
-        public DateTime SubmissionDate { get; set; }
+        public int Score { get; set; }
+        public DateTime SubmissionDate { get; set; }   
     }
 }
